backend/pyproject.toml
[project]
name = "vulntracker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "psycopg2-binary>=2.9",
    "django-cors-headers>=4.3",
    "gunicorn>=22.0",
    "requests>=2.31",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = "tests.py test_*.py"
pythonpath = "."
